refactor: split oversized source files
