+++ crates/clud-bin/src/session.rs
@@ -886,250 +886,15 @@ fn stdin_chunk_requests_interrupt(chunk: &[u8]) -> bool {
     chunk.contains(&0x03)
 }
 
-/// Bracketed-paste byte sequence emitted by xterm-class terminals when
-/// the user pastes (or drags-and-drops) text.
-const PASTE_START: &[u8] = b"\x1b[200~";
-const PASTE_END: &[u8] = b"\x1b[201~";
+mod bracketed_paste;
 
-/// Stream-resumable bracketed-paste detector that, on each completed
-/// paste, runs the buffered inner content through
-/// [`looks_like_dropped_path`] / [`normalize_dropped_path`] to canonicalize
-/// terminal-specific drop encodings (issue #63 / #79).
-///
-/// Behavior:
-/// - Bytes outside any paste pass through unchanged.
-/// - Bytes inside a bracketed paste are buffered. On `\x1b[201~`:
-///     - If `looks_like_dropped_path(inner)` returns true, emit
-///       `\x1b[200~` + `normalize_dropped_path(inner)` + `\x1b[201~`.
-///     - Otherwise, emit the original `\x1b[200~ inner \x1b[201~` verbatim.
-/// - The detector survives across chunks: a paste split across reads is
-///   reassembled correctly.
-///
-/// The PASS-IT-VERBATIM rule on non-path content is essential — a
-/// multi-line code paste must not be mutated, even if its first line
-/// happens to start with `/`.
-pub struct BracketedPasteNormalizer {
-    /// How many bytes of `PASTE_START` we've matched while outside a
-    /// paste. 0..PASTE_START.len().
-    start_match: usize,
-    /// `Some(buf)` while we are inside a paste body. The buffer holds
-    /// the *inner* paste content (no `\x1b[200~` prefix and no terminal
-    /// `\x1b[201~`).
-    inside: Option<Vec<u8>>,
-    /// How many bytes of `PASTE_END` we've matched while inside a paste.
-    end_match: usize,
-}
+pub use bracketed_paste::BracketedPasteNormalizer;
+#[cfg(test)]
+use bracketed_paste::{PASTE_END, PASTE_START};
 
-impl BracketedPasteNormalizer {
-    pub fn new() -> Self {
-        Self {
-            start_match: 0,
-            inside: None,
-            end_match: 0,
-        }
-    }
+mod interrupt;
 
-    /// Process a chunk, returning the byte stream that should be
-    /// forwarded downstream (PTY master, in production).
-    pub fn process(&mut self, chunk: &[u8]) -> Vec<u8> {
-        let mut out = Vec::with_capacity(chunk.len());
-        for &b in chunk {
-            if let Some(buf) = self.inside.as_mut() {
-                // We are inside a paste body. Look for PASTE_END.
-                if b == PASTE_END[self.end_match] {
-                    self.end_match += 1;
-                    if self.end_match == PASTE_END.len() {
-                        // Complete: emit normalized form and reset.
-                        let inner = std::mem::take(buf);
-                        self.inside = None;
-                        self.end_match = 0;
-                        emit_paste_block(&mut out, &inner);
-                    }
-                    continue;
-                }
-
-                // PASTE_END prefix broke. Flush any partial-end bytes
-                // back into the inner buffer, then this byte too.
-                if self.end_match > 0 {
-                    buf.extend_from_slice(&PASTE_END[..self.end_match]);
-                    // The byte that broke the prefix may itself start a
-                    // new PASTE_END match.
-                    self.end_match = if b == PASTE_END[0] { 1 } else { 0 };
-                    if self.end_match == 0 {
-                        buf.push(b);
-                    }
-                } else {
-                    buf.push(b);
-                }
-            } else {
-                // We are outside a paste. Look for PASTE_START.
-                if b == PASTE_START[self.start_match] {
-                    self.start_match += 1;
-                    if self.start_match == PASTE_START.len() {
-                        // Complete: enter paste body.
-                        self.start_match = 0;
-                        self.end_match = 0;
-                        self.inside = Some(Vec::new());
-                    }
-                    continue;
-                }
-
-                // PASTE_START prefix broke. Flush partial bytes verbatim.
-                if self.start_match > 0 {
-                    out.extend_from_slice(&PASTE_START[..self.start_match]);
-                    self.start_match = if b == PASTE_START[0] { 1 } else { 0 };
-                    if self.start_match == 0 {
-                        out.push(b);
-                    }
-                } else {
-                    out.push(b);
-                }
-            }
-        }
-        out
-    }
-}
-
-impl Default for BracketedPasteNormalizer {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Helper for `BracketedPasteNormalizer::process` — given a captured
-/// inner paste body, emit the wrapped (and possibly path-normalized)
-/// bracketed-paste block to `out`.
-fn emit_paste_block(out: &mut Vec<u8>, inner: &[u8]) {
-    out.extend_from_slice(PASTE_START);
-    // Decide path-rewrite on the WHOLE buffer, not per-line: a
-    // multi-line code paste with a path-shaped first line must remain
-    // verbatim. `looks_like_dropped_path` is conservative — it requires
-    // the entire trimmed string to look like a single path token.
-    let s = match std::str::from_utf8(inner) {
-        Ok(s) => s,
-        Err(_) => {
-            out.extend_from_slice(inner);
-            out.extend_from_slice(PASTE_END);
-            return;
-        }
-    };
-    if looks_like_dropped_path(s) {
-        let normalized = normalize_dropped_path(s);
-        out.extend_from_slice(normalized.as_bytes());
-    } else {
-        out.extend_from_slice(inner);
-    }
-    out.extend_from_slice(PASTE_END);
-}
-
-fn reap_pty_exit(process: &NativePtyProcess) -> i32 {
-    process.wait_impl(Some(1.0)).unwrap_or(1)
-}
-
-/// Escalate the `interrupted` flag to a real child-kill. Called once the
-/// pump has observed the flag. Platform-split because `send_interrupt_impl`
-/// is a byte-write on Windows (duplicates the 0x03 already forwarded via
-/// raw-mode stdin) and a pgroup-SIGINT on POSIX (cooperative, no duplicate).
-///
-/// Both branches now try a fire-and-forget daemon handoff first so the
-/// CLI can return to the shell in sub-100ms instead of blocking on
-/// `process.wait_impl(Some(2.0))` (POSIX) or the ConPTY close-event
-/// chain that occasionally lets cmd.exe surface the `Terminate batch job
-/// (Y/N)?` prompt (Windows). The daemon's background thread does the
-/// real kill_tree from out of the user's hot path.
-fn interrupt_pty_process(process: &NativePtyProcess, verbose: bool) -> i32 {
-    let pid = process.pid().ok().flatten();
-    let handed_off = match pid {
-        Some(pid) => match crate::daemon::default_state_dir() {
-            Ok(state_dir) => {
-                let ok = crate::daemon::try_handoff_kill_to_daemon(
-                    &state_dir,
-                    &[pid],
-                    Some("ctrl_c_pty"),
-                );
-                crate::ctrl_c_track::record_handoff(
-                    ok,
-                    Some(if ok {
-                        "ctrl_c_pty"
-                    } else {
-                        "daemon_unreachable"
-                    }),
-                );
-                ok
-            }
-            Err(_) => {
-                crate::ctrl_c_track::record_handoff(false, Some("no_state_dir"));
-                false
-            }
-        },
-        None => {
-            crate::ctrl_c_track::record_handoff(false, Some("no_child_pid"));
-            false
-        }
-    };
-
-    #[cfg(windows)]
-    {
-        if handed_off {
-            // Daemon owns the kill. Skip ConPTY close so cmd.exe doesn't
-            // get a chance to print "Terminate batch job (Y/N)?" before
-            // the Job Object reaps it on our process exit.
-            if verbose {
-                verbose_log::log("[clud] interrupted via Ctrl+C (pty, handed to daemon)");
-            }
-            return 130;
-        }
-        // Closing the PTY triggers ConPTY's CTRL_CLOSE_EVENT path and
-        // tears the child down without writing a second 0x03 byte.
-        let _ = process.close_impl();
-        if verbose {
-            verbose_log::log("[clud] interrupted via Ctrl+C (pty)");
-        }
-        130
-    }
-    #[cfg(not(windows))]
-    {
-        if handed_off {
-            // Daemon will SIGKILL the tree; skip the 2s blocking wait.
-            let _ = process.close_impl();
-            if verbose {
-                verbose_log::log("[clud] interrupted via Ctrl+C (pty, handed to daemon)");
-            }
-            return 130;
-        }
-        // Belt-and-braces: portable-pty's `send_interrupt` queries
-        // `tcgetpgrp(master_fd)` to find the FG pgroup and signals it.
-        // If that query returns None (no controlling-terminal coupling
-        // ever established, or the slave already lost FG), the library
-        // falls back to writing a raw 0x03 byte to the master — which
-        // only fires SIGINT if the slave still has ISIG set and the
-        // child is actively reading. Tests that drive a sleep-only
-        // child (issue #159) fail under that fallback because nothing
-        // converts 0x03 to a signal. Send through both paths and also
-        // signal the child's PID tree directly so all three vectors are
-        // covered. Then `close_impl` ensures the master is torn down so
-        // the pump loop doesn't keep polling a half-dead PTY.
-        let _ = process.send_interrupt_impl();
-        let tree_signal_result = process.terminate_tree_impl();
-        let _ = process.wait_impl(Some(2.0));
-        let _ = process.close_impl();
-        if verbose {
-            verbose_log::log("[clud] interrupted via Ctrl+C (pty)");
-            if let Err(err) = tree_signal_result {
-                verbose_log::log(format_args!(
-                    "[clud] pty interrupt: tree-signal fallback failed: {err}"
-                ));
-            }
-        }
-        // Match the Windows branch above: always report SIGINT's
-        // shell-convention 130 when clud itself handled the Ctrl-C.
-        // The child's actual exit code is observable via the wait_impl
-        // call above (stored in returncode for diagnostics); we just
-        // don't propagate it, because the contract is "user pressed
-        // Ctrl-C → exit 130" regardless of how the child shut down.
-        130
-    }
-}
+use interrupt::{interrupt_pty_process, reap_pty_exit};
 
 #[cfg(test)]
 #[path = "session_tests.rs"]
